Fix warning.
git-svn-id: https://fuse-emulator.svn.sourceforge.net/svnroot/fuse-emulator/trunk@3431 6c7d1b9a-d430-0410-8293-a2b96dacb39b

+++ fuse-utils/rzxtool.c
@@ -105,8 +105,8 @@ write_file( unsigned char *buffer, size_t length, const char *filename )
 
   bytes = fwrite( buffer, 1, length, f );
   if( bytes != length ) {
-    fprintf( stderr, "%s: wrote only %d of %d bytes to `%s'\n", progname,
-	     bytes, length, filename );
+    fprintf( stderr, "%s: wrote only %lu of %lu bytes to `%s'\n", progname,
+	     (unsigned long)bytes, (unsigned long)length, filename );
     return 1;
   }
